+++ src/main/java/com/tave/weathertago/service/auth/AuthServiceImpl.java
@@ -3,7 +3,7 @@ package com.tave.weathertago.service.auth;
 import com.tave.weathertago.apiPayload.code.status.ErrorStatus;
 import com.tave.weathertago.apiPayload.exception.handler.AuthHandler;
 import com.tave.weathertago.apiPayload.exception.handler.UserHandler;
-import com.tave.weathertago.config.security.jwt.JwtTokenProvider;
+import com.tave.weathertago.security.jwt.JwtTokenProvider;
 import com.tave.weathertago.converter.AuthConverter;
 import com.tave.weathertago.domain.User;
 import com.tave.weathertago.dto.Auth.AuthRequestDTO;
